+++ lib/format.rb
@@ -1,8 +1,13 @@
-require 'mechanize'
+require 'net/http'
+require 'nokogiri'
+
 class String
   def to_kanhira
-    agent = Mechanize.new
-    agent.get("https://yomikatawa.com/kanji/#{self}").search('#yomikata tbody tr td').first.inner_text
+    http = Net::HTTP.new('yomikatawa.com', 443)
+    http.use_ssl = true
+    path = File.join('/kanji', self)
+    html = http.get(path).body
+    Nokogiri::HTML(html).search('#yomikata tbody tr td').first.inner_text
   end
 
   def to_kana