+++ lib/miyabi.rb
@@ -1,6 +1,6 @@
-require_relative "miyabi/version"
-require_relative "format"
-require_relative "check"
+require_relative 'miyabi/version'
+require_relative 'format'
+require_relative 'check'
 
 module Miyabi
 end